backend/requirements.txt
fastapi>=0.103,<1.0
uvicorn[standard]>=0.22,<1.0
SQLAlchemy>=2.0,<3.0
pydantic>=2.0,<3.0
python-multipart>=0.0.9,<1.0
openai>=2.0,<3.0
PyYAML>=6.0,<7.0
pypdf>=5.0,<6.0
python-docx>=1.1,<2.0
PyMuPDF>=1.24,<2.0
rapidocr_onnxruntime>=1.3,<2.0
chromadb>=0.4.0,<1.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
